experiments/eips/requirements.txt
beautifulsoup4==4.13.4
datasets==3.5.0
langchain==0.3.23
langchain_chroma==0.2.3
langchain_huggingface==0.1.2
langgraph==0.3.31
peft==0.15.1
pydantic==2.11.3
PyGithub==2.6.1
PyGithub==2.6.1
python-dotenv==1.1.0
Requests==2.32.3
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3
